Choose to show empty sections' header or not.
Whilst v0.2.1 disabled showing headers for empty sections, this change
lets users to re-enable the old behaviour of showing empty sections
headers if they so choose.  The default behaviour is still to not show
them as per v0.2.1's change, but for applications that relied upon the
old behaviour, this is a much better solution than adding dummy items to
sections just to show headers.